Revert "websocket stuff wip"
This reverts commit 9f53e8b2
