--- examples/sse-with-state/.dockerignore
@@ -1,11 +1,0 @@
-# Project exclude paths
-/tmp/
-node_modules/
-dist/
-js/dist
-js/node_modules
-go.work
-go.work.sum
-.idea
-!framework/assets/dist
-__htmgo
