--- examples/sse-with-state/.gitignore
@@ -1,6 +1,0 @@
-/assets/dist
-tmp
-node_modules
-.idea
-__htmgo
-dist
