--- examples/sse-with-state/Taskfile.yml
@@ -1,20 +1,0 @@
-version: '3'
-
-tasks:
-  run:
-    cmds:
-      - go run github.com/maddalax/htmgo/cli/htmgo@latest run
-    silent: true
-
-  build:
-    cmds:
-      - go run github.com/maddalax/htmgo/cli/htmgo@latest build
-
-  docker:
-    cmds:
-      - docker build .
-
-  watch:
-    cmds:
-      - go run github.com/maddalax/htmgo/cli/htmgo@latest watch
-    silent: true
