--- examples/sse-with-state/go.mod
@@ -1,14 +1,0 @@
-module sse-with-state
-
-go 1.23.0
-
-require github.com/maddalax/htmgo/framework v0.0.0-20241006162137-150c87b4560b
-
-require (
-	github.com/go-chi/chi/v5 v5.1.0 // indirect
-	github.com/gobwas/httphead v0.1.0 // indirect
-	github.com/gobwas/pool v0.2.1 // indirect
-	github.com/gobwas/ws v1.4.0 // indirect
-	github.com/google/uuid v1.6.0 // indirect
-	golang.org/x/sys v0.6.0 // indirect
-)
