--- examples/sse-with-state/tailwind.config.js
@@ -1,5 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ["**/*.go"],
-  plugins: [],
-};
