+++ framework/assets/js/htmgo.ts
@@ -6,8 +6,7 @@ import "./htmxextensions/response-targets";
 import "./htmxextensions/mutation-error";
 import "./htmxextensions/livereload"
 import "./htmxextensions/htmgo";
-import "./htmxextensions/ws"
-import "./htmxextensions/ws-event-handler"
+import "./htmxextensions/sse"
 
 // @ts-ignore
 window.htmx = htmx;
