+++ framework/h/extensions.go
@@ -3,7 +3,7 @@ package h
 import "strings"
 
 func BaseExtensions() string {
-	extensions := []string{"path-deps", "response-targets", "mutation-error", "htmgo", "sse", "ws"}
+	extensions := []string{"path-deps", "response-targets", "mutation-error", "htmgo", "sse"}
 	if IsDevelopment() {
 		extensions = append(extensions, "livereload")
 	}
