+++ framework/h/tag.go
@@ -2,7 +2,6 @@ package h
 
 import (
 	"fmt"
-	"github.com/maddalax/htmgo/framework/internal/util"
 	"strconv"
 )
 
@@ -152,10 +151,6 @@ func Pre(children ...Ren) *Element {
 
 func Div(children ...Ren) *Element {
 	return Tag("div", children...)
-}
-
-func GenId() string {
-	return util.RandSeq(6)
 }
 
 func Article(children ...Ren) *Element {
